Adding link to Netlify deployment.
Note that the home.js vs. Home.js file is causing issues.  To work around, I cloned this project, deleted the .git folder and re-created a new git repository with the corrections (renaming file to Home.js).  Note that to get this to work, I had to run the npm build process locally - the build process on Netlify will fail due to this home.js issue.